Fix #1151: building a second colony will no longer trigger an OOS
CvUnitAI::AI_init can be called from the UI, so it can't use synced random.
Luckily the random number isn't needed because it's used to set birthmark.
Further down in the function, that int is set to the int argument.
In other words the random number isn't even used to begin with.