Added DesyncMinotor to the python interface
Went through practically all Cy files and made the pointer const.
They now have a function to make the pointer non-const and this function has an isSynced assert.
This means if python tries to write to game data, an assert will trigger if executed in async.

For this to work, a lot of functions in various Cv classes have been converted to const.

Also completely reorganized the Cy files in the project. They are now sorted by purpose.
Mostly there is a folder for each class, some folders are nested, like unit contains both
folders for CyUnit and CySelectionGroup.