Fix: adding natives will no longer place George Washington if the game fails to find a native civ
At game start when placing natives, if the game tries to add natives matching the requirements
of the game in question. Scenarios can have custom requirements.

If the game fails to find a native civ for a native player, the game now removes the player.
Before this bugfix, it would pick George Washington because he is first in the xml file.
This could result in multiple George Washingtons in a game, which then bugs the English king.